mp00 - extracts gene list to file
Can look at this for some stats regarding distribution of node degree,
or species types.
Also, the gene list will be useful later, when building the matrices
in mp02. To save space, only integers will be saved in the matrix files,
no text. A separate file will house the row & column names. These will
come from the genes known to be in the network.