mp02-2 - finished building matricies
For each edge type, builds a gene-gene matrix of paths, or edges.
This is stored in a compressed numpy array file, to quickly be
loaded as needed in the next step.
For edges that pass through non-gene nodes, the code converts them
to gene-gene edges and saves them in this array form.