added notation to use transpose matrices
If path A is the reverse of path B
(ie: A = a-b-c; B = c-b-a), then A and B can use
the same matrix, B will just use the transpose.
This cuts down a lot of duplicate data in memory.
Also, made minor changes to output
(ie: adding .txt to text files).