added functions to calculate the statistics
Added functions to: load the appropriate metapath
matrix, get the counts of a path within a sample,
get the mean and standard deviation of paths within
a set of random samples, and calculate the z-scores.
This is done with a priority on only loading each
matrix _once_, as this is a time-consuming step.