built new batch path ranking method
The slowest part of the prediction process is
reading in each path matrix. This method takes
a batch of samples, and only reads in each path
matrix once, performing the calculations for
each sample while that matrix is in memory.
Predicting/ranking connected genes from those
paths is left to another script (part 02b).